+++ package.json
@@ -13,7 +13,7 @@
     "@astrojs/check": "^0.9.1",
     "@astrojs/partytown": "^2.1.2",
     "@astrojs/vue": "^4.5.0",
-    "astro": "^4.13.1",
+    "astro": "^5.1.1",
     "typescript": "^5.5.4",
     "vue": "^3.4.37",
     "yarn": "^1.22.22"
